refactor(di): Move on_startup to a separate module
